[spinel] allow registering callback to handle compatibility errors (#10724)
When the RCP capabilities check fails, a mechanism should be provided
to notify the user of the error instead of causing the program to
crash.

For example, the hardware reboot function provided in the Spinel
driver interface could be invoked. The specific hardware reboot
implementation is left to the user, such as performing an RCP update.