[border-agent] invoke ephemeral key callback on session connection (#10699)
This commit updates the Border Agent to invoke the ephemeral key
callback when a commissioner candidate successfully establishes a
secure session using the ephemeral key. This can be identified by
checking the state `otBorderAgentGetState()` and matching it with
`OT_BORDER_AGENT_STATE_ACTIVE`. This new signal can be used to stop
advertising the mDNS service "_meshcop-e._udp" earlier (since the
ephemeral key is one-time use).