[secure-transport] use `mbedtls_ssl_is_handshake_over()` (#10973)
This commit updates how the handshake completion is checked on MBedTLS
version 3.0 or later. Instead of accessing the private `state`
variable, the `mbedtls_ssl_is_handshake_over()` function is used.
This follows the recommendation of the MbedTLS documentation to avoid
using the deprecated `mSsl->MBEDTLS_PRIVATE(state)` access on newer
versions.